+++ Handin3/Code/AirTrafficMonitor/Interfaces/FlightTracker/IFlightTrackerSingle.cs
@@ -12,9 +12,9 @@ namespace AirTrafficMonitor.Interfaces
     {
         event EventHandler<FlightTrackUpdatedEventArgs> FlightTrackUpdated;
         string GetTag();
-        float GetCurrentAltitude();
-        float GetCurrentVelocity();
-        float GetCurrentCourse();
+        double GetCurrentAltitude();
+        double GetCurrentVelocity();
+        double GetCurrentCourse();
         Vector2 GetCurrentPosition();
         void AddDataPoint(FTDataPoint dp);
         FTDataPoint GetNewestDataPoint();
